lib/requirements.txt
htools
incendio
openai
pyyaml==5.2.*
